refactor and add phpcheck.xml subscriber loading
