+++ src/Subscribers/ConsoleReporter.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Datashaman\PHPCheck\Reporters;
+namespace Datashaman\PHPCheck\Subscribers;
 
 use Datashaman\PHPCheck\CheckCommand;
 use Datashaman\PHPCheck\CheckEvents;
