+++ src/Subscribers/Reporter.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+/*
+ * This file is part of the phpcheck package.
+ *
+ * ©Marlin Forbes <marlinf@datashaman.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace Datashaman\PHPCheck\Subscribers;
+
+abstract class Reporter extends Subscriber
+{
+}
